requirements.txt
openai
langchain
python-dotenv
streamlit
PyPDF2
faiss-cpu
tiktoken
unstructured
"unstructured[pdf]"
